requirements.txt
Keras==2.2.4
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
pytest==4.0.1
wget==3.2
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
tensorflow-gpu==1.12.0
